Added temperature devices, contact to main driver some performance and bug fixes
Added ability to handle a zone with a Temperature Probe attached to the main board. Automatically assigned when
  the zone's description has the word "Temperature" in it or zone device can be manually changed.
Added the device type Keypad for temperature readings.
Both devices are assigned the Virtual Temperature Sensor driver.  This is an experimental feature.  Since the panel
  does not volunteer temperature data, it must be requested either manually for all devices using the Request Temperature
  Data button on the main driver or by setting up a rule to periodically execute the RequestTemperatureData command.
Added "ContactSensor" as a capability for HSM monitoring.  The contact will open if a Burglar Alarm or Police Alarm is triggered.
Changed the method of importing devices to greatly improve performance and reduce panel communication.
Fixed an issue not deleting child devices when the main driver is uninstalled.
Fixed an issue with the name when importing a device with the "Show On Keypad" flag set.