Added initial polling, more log settings and device reporting features.
Added polling of device status once connected or reconnected to the panel.
Added Enable debug logging and Enable descriptionText logging to the main device.  Debug is no longer set for the
  driver from within the application.  Info logging can now be turned on or off for the main device.
Added zone, output and task reporting so child devices can register to be updated via the report command when another
  device status has changed.
  I am using this for a door control driver assigned to an output that needs to be aware of the state of the contact
  attached to the door.