Major version release with many new additions.
Added keypad functionality such as chime and temperature measurement by assigning a keypad number to this device.  Device area is derived from the area the keypad is assigned to.
Added Function 1-6 keys as pushable buttons and retrieving of their description.
Added Function key LED status.
Added keypad beep, chime and chime mode status.
Added alarm and system trouble status.
Added request of zone voltage to be written to info log.
Added speak word/phrase using the Elk word or phrase number.
Added showTextOnKeypads to display a message on all keypads.
Added zone bypass and zone trigger commands.
Added setting if temperatures from the panel is in C or F.
Added setting to control if you want the hub's location set when the panel is armed/disarmed.
Added settings to control what arming or disarming happens when the switch assigned to this device is turned on or off.  Primarily for use by the dashboard or Google/Alexa.
Enhanced logging setting to allow control of whether Area or Keypad events get logged.
Added sendEmail event when an email is sent by the panel.
Added warning log when certain data is written to the panel log.
Added warning log on zone devices when a zone is in alarm.
Added status event on zone devices when a zone has a trouble code.
Added pushed status when certain keys are pressed a the keypad.
Added invalidUser status when an invalid user code is entered on a keypad.  This could be used to trigger rules based on a specific invalid user code being entered.
Added refresh if Elk RP programming occurs and warning log of Elk RP connections.
Added descriptions to events for clarity.
Added "arming" armStatus event when panel is armed and the exit delay has not yet expired.
Added calls to Elk application when arm/disarm and exit delay events occur for HSM, TTS and notifications.
Added warning log when alarm reporting occurs.
Added hooks for separate custom Elk Keypad and Speech drivers.
Changed task to be a momentary pushable button instead of a switch.